+++ app/Controller/CoPetitionsController.php
@@ -157,9 +157,7 @@ class CoPetitionsController extends StandardController {
     'processConfirmation'          => 'collectIdentifier',
     'collectIdentifier'            => 'checkEligibility',
     // approve is re-entry point following approval
-    'approve'                      => 'sendApprovalNotification',
-    'sendApprovalNotification'     => 'finalize',
-    'deny'                         => 'finalize',
+    'approve'                      => 'finalize',
     'finalize'                     => 'provision',
     'provision'                    => 'redirectOnConfirm'
   );
@@ -253,6 +251,22 @@ class CoPetitionsController extends StandardController {
     }
     
     if(!$this->request->is('restful')) {
+      if(!in_array($this->action, array('approve', 'finalize', 'provision', 'redirectOnConfirm', 'start', 'view'))) {
+        // If the petition is Finalized (or Denied/Declined), no further actions are permitted
+        // (except the post processing actions of provisioning and redirection). We also need
+        // to allow approve and finalize so plugins can run.
+        $status = $this->CoPetition->field('status', array('CoPetition.id' => $this->parseCoPetitionId()));
+
+        if($status && in_array($status, 
+                               array(PetitionStatusEnum::Declined,
+                                     PetitionStatusEnum::Denied,
+                                     PetitionStatusEnum::Duplicate,
+                                     PetitionStatusEnum::Finalized))) {
+          $this->Flash->set(_txt('er.pt.readonly'), array($status));
+          $this->redirect("/");
+        }
+      }
+
       // Under certain circumstances, we may wish to drop authentication.
       $noAuth = false;
       
@@ -673,17 +687,6 @@ class CoPetitionsController extends StandardController {
     
     // Make sure we don't issue a redirect
     return;
-  }
-  
-  /**
-   * Deny a petition.
-   *
-   * @since  COmanage Registry v0.5
-   * @param  Integer Petition ID
-   */
-  
-  public function deny($id) {
-    $this->dispatch('deny', $id);
   }
   
   /**
@@ -1244,15 +1247,33 @@ class CoPetitionsController extends StandardController {
   
   protected function execute_approve($id) {
     // Let any Exceptions pass through
+
+    // As of v3.2.0, execute_approve handles both approve and deny.
+    $action = PetitionStatusEnum::Denied;
+    $result = _txt('rs.pt.deny');
+    $comment = "";
+
+    if(!empty($this->request->data['action'])
+      && $this->request->data['action'] == _txt('op.approve')) {
+      $action = PetitionStatusEnum::Approved;
+      $result = _txt('rs.pt.approve');
+    }
+
+    if(!empty($this->request->data['CoPetition']['approver_comment'])) {
+      $comment = $this->request->data['CoPetition']['approver_comment'];
+    }
     
     $this->CoPetition->updateStatus($id,
-                                    PetitionStatusEnum::Approved,
-                                    $this->Session->read('Auth.User.co_person_id'));
+                                    $action,
+                                    $this->Session->read('Auth.User.co_person_id'),
+                                    $comment);
     
-    $this->Flash->set(_txt('rs.pt.approve'), array('key' => 'success'));
+    $this->Flash->set($result, array('key' => 'success'));
+
+    // Send out approval or denial notification, if configured
+    $this->CoPetition->sendApprovalNotification($id, $this->Session->read('Auth.User.co_person_id'));
     
     // The step is done
-    
     $this->redirect($this->generateDoneRedirect('approve', $id));
   }
   
@@ -1326,29 +1347,7 @@ class CoPetitionsController extends StandardController {
     
     $this->redirect($this->generateDoneRedirect('collectIdentifier', $id));
   }
-  
-  /**
-   * Execute CO Petition 'deny' step
-   *
-   * @since  COmanage Registry v0.9.4
-   * @param Integer $id CO Petition ID
-   * @throws Exception
-   */
-  
-  protected function execute_deny($id) {
-    // Let any Exceptions pass through
-    
-    $this->CoPetition->updateStatus($id,
-                                    PetitionStatusEnum::Denied,
-                                    $this->Session->read('Auth.User.co_person_id'));
-    
-    $this->Flash->set(_txt('rs.pt.deny'), array('key' => 'success'));
-    
-    // The step is done
-    
-    $this->redirect($this->generateDoneRedirect('deny', $id));
-  }
-  
+
   /**
    * Execute CO Petition 'finalize' step
    *
@@ -1852,23 +1851,7 @@ class CoPetitionsController extends StandardController {
     // We're done, complete the step
     $this->redirect($this->generateDoneRedirect('selectOrgIdentity', $id));
   }
-  
-  /**
-   * Execute CO Petition 'sendApprovalNotification' step
-   *
-   * @since  COmanage Registry v0.9.4
-   * @param Integer $id CO Petition ID
-   * @throws Exception
-   */
-  
-  protected function execute_sendApprovalNotification($id) {
-    $this->CoPetition->sendApprovalNotification($id, $this->Session->read('Auth.User.co_person_id'));
-    
-    // The step is done
-    
-    $this->redirect($this->generateDoneRedirect('sendApprovalNotification', $id));
-  }
-  
+
   /**
    * Execute CO Petition 'sendApproverNotification' step
    *
@@ -2262,8 +2245,6 @@ class CoPetitionsController extends StandardController {
       }
       // Actual approval is handled by the approver
       $p['approve'] = $isApprover;
-      $p['deny'] = $isApprover;
-      $p['sendApprovalNotification'] = $isApprover;
       // Finalize and finalize steps could be reached by anyone, in theory
       foreach(array('finalize', 'provision') as $xstep) {
         switch($steps[$xstep]['role']) {
@@ -2764,17 +2745,6 @@ class CoPetitionsController extends StandardController {
   }
   
   /**
-   * Send approval notification for a new CO Petition
-   *
-   * @since  COmanage Registry v0.9.4
-   * @param  Integer $id CO Petition ID
-   */
-  
-  public function sendApprovalNotification($id) {
-    $this->dispatch('sendApprovalNotification', $id);
-  }
-  
-    /**
    * Send approver notification for a new CO Petition
    *
    * @since  COmanage Registry v0.9.4